+++ Packages/dev.ameye.linework/Runtime/SoftOutline/SoftOutlineUtils.cs
@@ -74,6 +74,7 @@ namespace Linework.SoftOutline
     {
         public const string AlphaCutout = "ALPHA_CUTOUT";
         public const string HardOutline = "HARD_OUTLINE";
+        public const string ScaleWithResolution = "SCALE_WITH_RESOLUTION";
     }
 
     static class Buffer