+++ Packages/dev.ameye.linework/CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.4] - 2024-11-13
+
+### Added
+
+- Added scale-with-resolution option for soft outline resulting in better performance at higher resolutions
+
+### Fixed
+
+- Fixed graphics format not being supported on Android
+
+## [1.2.3] - 2024-11-09
+
+### Fixed
+
+- Fixed edge detection not rendering on Unity 6000.0.22f1 or newer and Unity 2022.3.49f1 or newer
+
 ## [1.2.2] - 2024-11-06
 
 ### Fixed